added some threading to beam collision.(perf) changed the course beam collision to use ray and bounding box collision. This should be a small loop to check for hits.(experimental) Reverted cic and bridge control values to 5 and 1. (balancing perf)
--HG--
branch : develop

+++ Ship_Game/Gameplay/Projectile.cs
@@ -994,7 +994,7 @@ namespace Ship_Game.Gameplay
                     this.inFlight.Play();
                 }
                 this.particleDelay -= elapsedTime;
-                if ((double)this.duration > 0.0)
+                if (this.duration > 0f)
                 {
                     this.duration -= elapsedTime;
                     if ((double)this.duration < 0.0)
